% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_all_msgs}
\alias{check_all_msgs}
\title{Output all messages with timestamps}
\usage{
check_all_msgs(data, ReturnData = FALSE)
}
\arguments{
\item{data}{A data table object output by \code{\link{prep_data}},
\code{\link{align_msg}}, or \code{\link{create_time_series}}.}

\item{ReturnData}{A logical indicating whether to return a data table 
containing Message Time information for each event.}
}
\value{
All Sample Messages found in the data.
}
\description{
\code{check_all_msgs} outputs all sample messages present in the data. 
Optionally, the result can be output to a dataframe containing all event-level
information.
}
\examples{
\dontrun{
library(VWPre)
# Check all messages in the data...
check_all_msgs(data = dat)
}
}
