% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_eye_recording}
\alias{check_eye_recording}
\title{Check which eyes were recorded during the experiment}
\usage{
check_eye_recording(data)
}
\arguments{
\item{data}{A data table object output by \code{\link{create_time_series}}.}
}
\value{
Text feedback and instruction.
}
\description{
\code{check_eye_recording} quickly checks if the dataset contains gaze data
in both the Right and Left interest area columns. It prints a summary and 
suggests which setting to use for the \code{Recording} parameter in the 
function \code{\link{select_recorded_eye}}.
}
\examples{
\dontrun{
library(VWPre)
# Create a unified columns for the gaze data...
check_eye_recording(dat)
}
}
