% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_ia}
\alias{check_ia}
\title{Check the interest area IDs and labels}
\usage{
check_ia(data)
}
\arguments{
\item{data}{A data table object output by \code{\link{relabel_na}}.}
}
\value{
The value(s) and label(s) of interest areas and how they map for
each eye.
}
\description{
\code{check_ia} examines both the interest area IDs and interest area labels
(and their mapping) for both eyes. It returns a summary of the information
and will stop if the data are not consistent with the requirements for
further processing.
}
\examples{
\dontrun{
library(VWPre)
# Check the interest area information...
check_ia(dat)
}
}
