% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_samples_per_bin}
\alias{check_samples_per_bin}
\title{Check the number of samples in each bin}
\usage{
check_samples_per_bin(data = data)
}
\arguments{
\item{data}{A data table object output by \code{\link{bin_prop}}.}
}
\value{
A printed summary of the number of samples in each bin.
}
\description{
\code{check_samples_per_bin} determines the number of samples in each
bin produced by \code{\link{bin_prop}}.
This function may be helpful for determining the obligatory parameter 
`ObsPerBin` which is input to \code{\link{transform_to_elogit}}.
}
\examples{
\dontrun{
library(VWPre)
# Determine the number of samples per bin...
check_samples_per_bin(dat)
}
}

