% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_samplingrate}
\alias{check_samplingrate}
\title{Determine the sampling rate present in the data}
\usage{
check_samplingrate(data, ReturnData = FALSE)
}
\arguments{
\item{data}{A data table object output by \code{\link{select_recorded_eye}}.}

\item{ReturnData}{A logical indicating whether to return a data table containing 
a new column called SamplingRate}
}
\value{
A printed summary and/or a data table object
}
\description{
\code{check_samplingrate} determines the sampling rate in the data.
This function is helpful for determining the obligatory parameter input to 
\code{\link{bin_prop}}. If different sampling rates were used, the 
function adds a sampling rate column, which can be used to subset the 
data for further processing.
}
\examples{
\dontrun{
library(VWPre)
# Determine the sampling rate...
check_samplingrate(dat)
}
}
