% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{create_time_series}
\alias{create_time_series}
\title{Create a meaningful time series column}
\usage{
create_time_series(data = data, Offset = Offset)
}
\arguments{
\item{data}{A data table object output by \code{\link{relabel_na}}.}

\item{Offset}{A positive integer indicating amount of time in milliseconds.}
}
\value{
A data table object.
}
\description{
\code{create_time_series} aligns the starting point for each 
event, creates a time series for each event including the offset for the 
amount of time prior to the onset auditory stimulus. The time series
is indicated in a new column called Time.
}
\examples{
\dontrun{
library(VWPre)
# To create the Time column...
df <- create_time_series(data = dat, Offset = 100)
}
}

