% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{.check_for_PupilPre}
\alias{.check_for_PupilPre}
\title{Internal helper function, not intended to be called externally.}
\usage{
.check_for_PupilPre(type, suggest)
}
\arguments{
\item{type}{A string, "NotAvailable" of "UseOther".}

\item{suggest}{A string, PupilPre function name.}
}
\value{
Text feedback and instruction.
}
\description{
Internal helper function, not intended to be called externally.
}
