% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_indiv_app}
\alias{plot_indiv_app}
\title{Plots diagnostic average plots of subjects/items.}
\usage{
plot_indiv_app(data = data)
}
\arguments{
\item{data}{A data table object output by either \code{\link{bin_prop}}. 
\code{\link{transform_to_elogit}}, or \code{\link{create_binomial}}.}
}
\description{
\code{plot_indiv_app} calculates and plots interest area averages for a 
given subject/item.
}
\examples{
\dontrun{
library(VWPre)
# For plotting the grand average with the included theme
plot_indiv_app(data = dat)
} 
}

