% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_transformation_app}
\alias{plot_transformation_app}
\title{Plots diagnostic plots of the empirical logit transformation.}
\usage{
plot_transformation_app()
}
\description{
\code{plot_transformation_app} plots the empirical logit values for a 
given number of observations and constant against proportions, in order
to examine the effect of these variables on the resulting transformation.
}
\examples{
\dontrun{
library(VWPre)
# For plotting the empirical logit transformation
plot_transformation_app() 
}
}
