% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.r
\docType{data}
\name{VWdat}
\alias{VWdat}
\title{This is a sample eye-tracking dataset included in the package}
\description{
This is a sample eye-tracking dataset included in the package
}
\author{
Vincent Porretta
}
\keyword{data}
